+++ app/src/main/java/com/example/Dietapp/categorias/Fruta.java
@@ -21,8 +21,8 @@ import com.example.myapplicationfinal.R;
 import java.util.ArrayList;
 
 public class Fruta extends AppCompatActivity {
-    double resultado,resultado2,resultado3,resultado4,resultado5,resultado6;
-    double total;
+    float resultado,resultado2,resultado3,resultado4,resultado5,resultado6;
+    float total;
     private Spinner sp;
     private EditText edi;
     TextView text;
@@ -133,7 +133,7 @@ public class Fruta extends AppCompatActivity {
         toastNada3.show();
         SharedPreferences myPreferences = PreferenceManager.getDefaultSharedPreferences(Fruta.this);
         SharedPreferences.Editor myEditor = myPreferences.edit();
-        myEditor.putInt("fru", (int) total);
+        myEditor.putFloat("fru", (int) total);
         myEditor.commit();
     }
